Refactored buildRecendModelRevision to use a child process
